apps/learn-documents/pyproject.toml
[tool.poetry]
name = "learn-documents"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "learn_documents"

  [tool.poetry.dependencies]
  python = ">=3.10,<=3.12"
  llama-index-core = "^0.10.53.post1"
  llama-index-llms-openai = "^0.1.25"
  llama-index-readers-file = "^0.1.29"
  chromadb = "^0.5.3"
  llama-index-vector-stores-chroma = "^0.1.10"
  llama-index-embeddings-openai = "^0.1.10"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
